♻️ Fix all broken tests caused by the migration of the external service *Slack* checking point abstraction.
