+++ test/e2e_test/scheduler/checking_points/test_checking_points_e2e.py
@@ -6,9 +6,9 @@ import pytest
 
 from gearmeshing_ai.scheduler.checking_points.base import (
     CheckingPointType,
-    SlackCheckingPoint,
 )
 from gearmeshing_ai.scheduler.checking_points.clickup.base import ClickUpCheckingPoint
+from gearmeshing_ai.scheduler.checking_points.slack.base import SlackCheckingPoint
 from gearmeshing_ai.scheduler.checking_points.registry import CheckingPointRegistry
 from gearmeshing_ai.scheduler.models.checking_point import CheckResult
 from gearmeshing_ai.scheduler.models.monitoring import MonitoringData, MonitoringDataType
